requirements.txt
absl-py==0.9.0
asgiref==3.2.3
astor==0.8.1
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
decorator==4.4.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.3
django-filter==2.2.0
django-heroku==0.3.1
django-rest-framework==0.1.0
djangorestframework==3.11.0
entrypoints==0.3
gast==0.3.3
google-pasta==0.1.8
grpcio==1.27.1
gunicorn==20.0.4
h5py==2.10.0
hatesonar==0.0.4
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.16.0
Jinja2==2.11.1
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.1.0
jupyter-core==4.6.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.2
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.4
nltk==3.4.5
notebook==6.0.3
numpy==1.18.1
pandas==1.0.1
pandocfilters==1.4.2
parso==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
pkg-resources==0.0.0
prometheus-client==0.7.1
prompt-toolkit==3.0.3
protobuf==3.11.3
psycopg2==2.8.4
ptyprocess==0.6.0
Pygments==2.5.2
pyrsistent==0.15.7
python-dateutil==2.8.1
pytz==2019.3
pyzmq==18.1.1
qtconsole==4.6.0
scikit-learn==0.22.1
scipy==1.4.1
Send2Trash==1.5.0
six==1.14.0
sqlparse==0.3.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
textblob==0.15.3
tornado==6.0.3
tqdm==4.42.1
traitlets==4.3.3
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.0
whitenoise==5.0.1
widgetsnbextension==3.5.1
wrapt==1.11.2
xgboost==0.90
zipp==2.2.0
